refactor: reorganize frontend into feature modules
