--- frontend/src/App.jsx
@@ -1,7 +1,0 @@
-import React from 'react'
-import GameUI from './components/Play/GameUI/GameUI.jsx'
-
-export default function App() {
-  return <GameUI />
-}
-
