+++ frontend/src/app/App.jsx
@@ -1,0 +1,8 @@
+import React from 'react'
+import GameUI from '../features/play/ui/GameUI.jsx'
+import '../shared/styles/theme.css'
+
+export default function App() {
+  return <GameUI />
+}
+
