--- frontend/src/components/Play/GameUI/GameUI.module.css
@@ -1,41 +1,0 @@
-:root { --font-size-player-name: 24px; --font-size-player-clock: 30px; --font-size-dagger-token: 14px; --font-size-captured-piece: 18px; --font-size-board-square: 12px; --font-size-board-piece: 14px; --font-size-stash-slot: 12px; --font-size-stash-piece: 12px; --font-size-action-button: 16px; }
-
-.viewport { width: 100vw; height: 100vh; display: flex; align-items: center; justify-content: center; background: #1a2f1f; }
-
-.playArea { position: relative; overflow: hidden; background: transparent; }
-
-.gameState {
-  position: absolute;
-  left: 0;
-  right: 0;
-}
-
-/* Action buttons */
-.actionButtons { position: absolute; display: flex; justify-content: space-between; background: transparent; }
-
-.actionButton {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  font-size: calc(var(--font-size-action-button) * 1.5);
-  font-weight: bold;
-  color: #ffffff;
-  background: #4815be;
-  border: 2px solid #fbbf24;
-  border-radius: 4px;
-  cursor: pointer;
-  transition: all 0.2s ease;
-}
-
-.actionButton:hover {
-  background: #7c3aed;
-  border-color: #f59e0b;
-  transform: translateY(-2px);
-  box-shadow: 0 4px 8px rgba(0,0,0,0.2);
-}
-
-.actionButton:active { transform: translateY(0); box-shadow: 0 2px 4px rgba(0,0,0,0.2); }
-
-.actionButton.bomb { background: #850202; }
-.actionButton.bomb:hover { background: #b91c1c; border-color: #f59e0b; }
-
