--- frontend/src/components/Play/GameUI/index.js
@@ -1,4 +1,0 @@
-// Export GameUI components
-export { default as PlayArea } from './PlayArea';
-export { default as PlayAreaDemo } from './PlayAreaDemo';
-export { default as GameUI } from './GameUI';
