--- frontend/src/components/Play/GameUI/parts/BackgroundDecor.jsx
@@ -1,3 +1,0 @@
-export default function BackgroundDecor() { return null }
-
-
