+++ frontend/src/features/play/assets/constants.js
@@ -1,0 +1,14 @@
+export const BOARD_ROWS = 6
+export const BOARD_COLS = 5
+export const STASH_ROWS = 2
+export const STASH_COLS = 5
+
+export const ACTIONS = {
+  MOVE: 1,
+  CHALLENGE: 2,
+  BOMB: 3,
+}
+
+export const MOVE_STATES = {
+  PENDING: 0,
+}
