+++ frontend/src/features/play/assets/pieces.js
@@ -1,0 +1,26 @@
+export const PIECE_IMAGES = {
+  0: {
+    0: '/assets/images/Pieces/PawnWhiteUnknown.svg',
+    1: '/assets/images/Pieces/PawnBlackUnknown.svg',
+  },
+  1: {
+    0: '/assets/images/Pieces/PawnWhiteKing.svg',
+    1: '/assets/images/Pieces/PawnBlackKing.svg',
+  },
+  2: {
+    0: '/assets/images/Pieces/PawnWhiteBomb.svg',
+    1: '/assets/images/Pieces/PawnBlackBomb.svg',
+  },
+  3: {
+    0: '/assets/images/Pieces/PawnWhiteBishop.svg',
+    1: '/assets/images/Pieces/PawnBlackBishop.svg',
+  },
+  4: {
+    0: '/assets/images/Pieces/PawnWhiteRook.svg',
+    1: '/assets/images/Pieces/PawnBlackRook.svg',
+  },
+  5: {
+    0: '/assets/images/Pieces/PawnWhiteKnight.svg',
+    1: '/assets/images/Pieces/PawnBlackKnight.svg',
+  },
+}
