+++ frontend/src/features/play/ui/Board.jsx
