+++ frontend/src/features/play/ui/GameUI.module.css
@@ -1,0 +1,9 @@
+.viewport { width: 100vw; height: 100vh; display: flex; align-items: center; justify-content: center; background: #1a2f1f; }
+
+.playArea { position: relative; overflow: hidden; background: transparent; }
+
+.gameState {
+  position: absolute;
+  left: 0;
+  right: 0;
+}
