+++ frontend/src/features/play/ui/PlayArea.jsx
