+++ frontend/src/features/play/ui/Stash.jsx
