+++ frontend/src/features/play/ui/index.js
@@ -1,0 +1,8 @@
+export { default as GameUI } from './GameUI.jsx'
+export { default as Board } from './Board.jsx'
+export { default as PlayerState } from './PlayerState.jsx'
+export { default as Stash } from './Stash.jsx'
+export { default as ActionButtons } from './ActionButtons.jsx'
+export { default as Queuer } from './Queuer.jsx'
+export { default as MatchFoundBanner } from './MatchFoundBanner.jsx'
+export { default as PlayArea } from './PlayArea.jsx'
