+++ frontend/src/main.jsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import { createRoot } from 'react-dom/client'
-import App from './App.jsx'
+import App from './app/App.jsx'
 
 const root = createRoot(document.getElementById('root'))
 root.render(
