+++ frontend/src/shared/ui/OverlayCard.jsx
@@ -1,0 +1,10 @@
+import React from 'react'
+import styles from './OverlayCard.module.css'
+
+export default function OverlayCard({ children }) {
+  return (
+    <div className={styles.overlay}>
+      <div className={styles.card}>{children}</div>
+    </div>
+  )
+}
