+++ frontend/src/shared/ui/OverlayCard.module.css
@@ -1,0 +1,23 @@
+.overlay {
+  position: absolute;
+  inset: 0;
+  background: rgba(0,0,0,0.55);
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  z-index: 1000;
+  pointer-events: none;
+}
+.card {
+  pointer-events: auto;
+  width: 100%;
+  max-width: 100%;
+  height: 160px;
+  padding: 18px 26px;
+  border-top: 2px solid #fbbf24;
+  border-bottom: 2px solid #fbbf24;
+  background: #4c1d95;
+  color: #ffffff;
+  box-shadow: 0 10px 30px rgba(0,0,0,0.35);
+  text-align: center;
+}
